I think I fixed the random generator

+++ webapp/src/main/java/edu/gatech/cs2340/risk/model/GameLogic.java
@@ -117,7 +117,7 @@ public class GameLogic
 
     public int rollDice(){
         Random gen = new Random();
-        int roll = gen.nextInt(6)+1;
+        int roll = (int) gen.nextInt(6)+1;
         return roll;
     }
